Renamed validate_single_root_transform
